+++ core/src/main/java/org/polypheny/db/schema/impl/AbstractEntity.java
@@ -106,7 +106,7 @@ public abstract class AbstractEntity implements Entity, Wrapper {
         if ( id == null ) {
             return Statistics.UNKNOWN;
         }
-        Integer rowCount = StatisticsManager.getInstance().rowCountPerTable( id );
+        Long rowCount = StatisticsManager.getInstance().rowCountPerTable( id );
 
         if ( rowCount == null ) {
             return Statistics.UNKNOWN;
