+++ webui/src/main/java/org/polypheny/db/webui/models/results/Result.java
@@ -90,7 +90,7 @@ public abstract class Result<E, F> {
     /**
      * Number of affected rows
      */
-    public int affectedTuples;
+    public long affectedTuples;
 
 
     /**
